requirements.txt
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
line-bot-sdk==1.10.0
psycopg2-binary==2.8.3